Add more trajectories
