+++ src/main/deploy/choreo/project.chor
@@ -237,6 +237,48 @@
      "val":2.0943951023931953
     }
    },
+   "M2_SL3_C":{
+    "x":{
+     "exp":"2.223389148712158 m",
+     "val":2.223389148712158
+    },
+    "y":{
+     "exp":"6.145318031311035 m",
+     "val":6.145318031311035
+    },
+    "heading":{
+     "exp":"0 deg",
+     "val":0.0
+    }
+   },
+   "M_SL3_C":{
+    "x":{
+     "exp":"2.7632617950439453 m",
+     "val":2.7632617950439453
+    },
+    "y":{
+     "exp":"3.3784706592559814 m",
+     "val":3.3784706592559814
+    },
+    "heading":{
+     "exp":"0 deg",
+     "val":0.0
+    }
+   },
+   "M_SL3_H":{
+    "x":{
+     "exp":"5.7876362800598145 m",
+     "val":5.7876362800598145
+    },
+    "y":{
+     "exp":"5.262604713439941 m",
+     "val":5.262604713439941
+    },
+    "heading":{
+     "exp":"0 deg",
+     "val":0.0
+    }
+   },
    "REEF":{
     "x":{
      "exp":"4.491799831390381 m",
@@ -305,6 +347,48 @@
     "heading":{
      "exp":"-SL3.heading",
      "val":-2.199114857512855
+    }
+   },
+   "STE":{
+    "x":{
+     "exp":"STJ.x",
+     "val":7.1
+    },
+    "y":{
+     "exp":"E.y",
+     "val":2.6014442565299394
+    },
+    "heading":{
+     "exp":"E.heading",
+     "val":5.235987755982989
+    }
+   },
+   "STF":{
+    "x":{
+     "exp":"STI.x",
+     "val":7.1
+    },
+    "y":{
+     "exp":"F.y",
+     "val":2.765768540129793
+    },
+    "heading":{
+     "exp":"F.heading",
+     "val":5.235987755982989
+    }
+   },
+   "STG":{
+    "x":{
+     "exp":"STH.x",
+     "val":7.2
+    },
+    "y":{
+     "exp":"G.y",
+     "val":3.872436782836916
+    },
+    "heading":{
+     "exp":"G.heading",
+     "val":0.0
     }
    },
    "STH":{
