+++ src/main/java/frc/robot/POI.java
@@ -1,0 +1,87 @@
+
+package frc.robot;
+
+import choreo.Choreo;
+import choreo.auto.AutoRoutine;
+import choreo.auto.AutoTrajectory;
+import edu.wpi.first.math.geometry.Pose2d;
+import frc.robot.util.AllianceFlipUtil;
+import frc.robot.util.ChoreoVariables;
+
+import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Optional;
+
+public enum POI {
+    A(true),
+    B(true),
+    C(true),
+    D(true),
+    E(true),
+    F(true),
+    G(true),
+    H(true),
+    I(true),
+    J(true),
+    K(true),
+    L(true),
+    // Intake left, right
+    SL3(false),
+    SR3(false),
+    // Starting for corresponding pose
+    STE(false),
+    STF(false),
+    STG(false),
+    STH(false),
+    STJ(false),
+    STI(false);
+
+    public final Pose2d bluePose;
+    public final Pose2d redPose;
+    public final boolean isReef;
+
+    private static List<String> TRAJECTORIES = List.of(Choreo.availableTrajectories());
+
+    private POI(boolean isReef) {
+        this.isReef = isReef;
+        bluePose = ChoreoVariables.getPose(this.name());
+        redPose = AllianceFlipUtil.flip(bluePose);
+    }
+
+    private POI(boolean isReef, Pose2d bluePose) {
+        this.isReef = isReef;
+        this.bluePose = bluePose;
+        redPose = AllianceFlipUtil.flip(bluePose);
+    }
+
+    public Pose2d flippedPose() {
+        return AllianceFlipUtil.shouldFlip() ? redPose : bluePose;
+    }
+
+    public Optional<AutoTrajectory> to(POI next, AutoRoutine routine) {
+        String name;
+        boolean reverse;
+        if (this.isReef && !next.isReef) {
+            name = next.name() + "-" + this.name();
+            if (!TRAJECTORIES.contains(name)) {
+                return Optional.empty();
+            }
+            reverse = true;
+        } else if (!this.isReef && next.isReef) {
+            name = this.name() + "-" + next.name();
+            reverse = false;
+        } else {
+            return Optional.empty();
+        }
+        return Optional.of(routine.trajectory(name, reverse ? 1 : 0));
+    }
+
+    public Optional<AutoTrajectory> toChecked(POI to, AutoRoutine routine) {
+        var option = to(to, routine);
+        if (option.isEmpty()) {
+
+            throw new NoSuchElementException(String.format("Missing %s-%s", this.name(), to.name()));
+        }
+        return option;
+    }
+}
